+++ kikuchipy/conftest.py
@@ -17,26 +17,90 @@
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
 import gc
+from numbers import Number
 import os
+from packaging import version
 import tempfile
 from typing import Tuple
 
 from diffpy.structure import Atom, Lattice, Structure
 from diffsims.crystallography import ReciprocalLatticePoint
+from hyperspy import __version__ as hs_version
+from hyperspy.misc.utils import DictionaryTreeBrowser
+import matplotlib.pyplot as plt
 import numpy as np
 from orix.crystal_map import CrystalMap, Phase, PhaseList
 from orix.quaternion.rotation import Rotation
 from orix.vector import Vector3d, neo_euler
 import pytest
 
-from kikuchipy.detectors import EBSDDetector
-from kikuchipy.generators import EBSDSimulationGenerator
+import kikuchipy as kp
 from kikuchipy.projections.ebsd_projections import (
     detector2reciprocal_lattice,
     detector2direct_lattice,
 )
-from kikuchipy.signals import EBSD
-from kikuchipy.simulations.features import KikuchiBand, ZoneAxis
+
+
+# ------------------------- Helper functions ------------------------- #
+
+
+def assert_dictionary(dict1, dict2):
+    """Assert that two dictionaries are (almost) equal.
+
+    Used to compare signal's axes managers or metadata in tests.
+    """
+    if isinstance(dict1, DictionaryTreeBrowser):
+        dict1 = dict1.as_dictionary()
+        dict2 = dict2.as_dictionary()
+    for key in dict2.keys():
+        if key in ["is_binned", "binned"] and version.parse(
+            hs_version
+        ) > version.parse(
+            "1.6.2"
+        ):  # pragma: no cover
+            continue
+        if isinstance(dict2[key], dict):
+            assert_dictionary(dict1[key], dict2[key])
+        else:
+            if isinstance(dict2[key], list) or isinstance(dict1[key], list):
+                dict2[key] = np.array(dict2[key])
+                dict1[key] = np.array(dict1[key])
+            if isinstance(dict2[key], (np.ndarray, Number)):
+                assert np.allclose(dict1[key], dict2[key])
+            else:
+                assert dict1[key] == dict2[key]
+
+
+def _get_spatial_array_dicts(
+    nav_shape: Tuple[int, int], step_sizes: Tuple[int, int] = (1.5, 1)
+) -> Tuple[dict, int]:
+    ny, nx = nav_shape
+    dy, dx = step_sizes
+    d = {"x": None, "y": None, "z": None}
+    map_size = 1
+    if nx > 1:
+        if ny > 1:
+            d["x"] = np.tile(np.arange(nx) * dx, ny)
+        else:
+            d["x"] = np.arange(nx) * dx
+        map_size *= nx
+    if ny > 1:
+        if nx > 1:
+            d["y"] = np.sort(np.tile(np.arange(ny) * dy, nx))
+        else:
+            d["y"] = np.arange(ny) * dy
+        map_size *= ny
+    return d, map_size
+
+
+# ------------------------------ Setup ------------------------------ #
+
+
+def pytest_sessionstart(session):  # pragma: no cover
+    _ = kp.data.nickel_ebsd_large(allow_download=True)
+
+
+# ----------------------------- Fixtures ----------------------------- #
 
 
 @pytest.fixture
@@ -56,7 +120,7 @@ def dummy_signal():
         dtype=np.uint8
     ).reshape((3, 3, 3, 3))
     # fmt: on
-    return EBSD(dummy_array)
+    return kp.signals.EBSD(dummy_array)
 
 
 @pytest.fixture
@@ -115,7 +179,7 @@ def pc1():
 @pytest.fixture(params=[(1,)])
 def detector(request, pc1):
     """A NORDIF UF1100 EBSD detector with a TSL PC."""
-    return EBSDDetector(
+    return kp.detectors.EBSDDetector(
         shape=(60, 60),
         binning=8,
         px_size=70,
@@ -183,7 +247,7 @@ def nickel_ebsd_simulation_generator(
     """Generator for EBSD simulations of Kikuchi bands for the Nickel
     data set referenced above.
     """
-    return EBSDSimulationGenerator(
+    return kp.generators.EBSDSimulationGenerator(
         detector=detector, phase=nickel_phase, rotations=nickel_rotations,
     )
 
@@ -197,7 +261,7 @@ def nickel_kikuchi_band(nickel_rlp, nickel_rotations, pc1):
 
     nav_shape = (5, 5)
 
-    detector = EBSDDetector(
+    detector = kp.detectors.EBSDDetector(
         shape=(60, 60),
         binning=8,
         px_size=70,
@@ -231,7 +295,7 @@ def nickel_kikuchi_band(nickel_rlp, nickel_rotations, pc1):
         hkl_detector[is_in_some_pattern], source=0, destination=nav_dim
     )
 
-    return KikuchiBand(
+    return kp.simulations.features.KikuchiBand(
         phase=phase,
         hkl=hkl,
         hkl_detector=hkl_detector,
@@ -248,7 +312,7 @@ def nickel_zone_axes(nickel_kikuchi_band, nickel_rotations, pc1):
 
     nav_shape = (5, 5)
 
-    detector = EBSDDetector(
+    detector = kp.detectors.EBSDDetector(
         shape=(60, 60),
         binning=8,
         px_size=70,
@@ -284,7 +348,7 @@ def nickel_zone_axes(nickel_kikuchi_band, nickel_rotations, pc1):
         uvw_detector[is_in_some_pattern], source=0, destination=nav_dim
     )
 
-    return ZoneAxis(
+    return kp.simulations.features.ZoneAxis(
         phase=phase,
         uvw=uvw,
         uvw_detector=uvw_detector,
@@ -327,23 +391,25 @@ def get_single_phase_xmap(rotations):
     return _get_single_phase_xmap
 
 
-def _get_spatial_array_dicts(
-    nav_shape: Tuple[int, int], step_sizes: Tuple[int, int] = (1.5, 1)
-) -> Tuple[dict, int]:
-    ny, nx = nav_shape
-    dy, dx = step_sizes
-    d = {"x": None, "y": None, "z": None}
-    map_size = 1
-    if nx > 1:
-        if ny > 1:
-            d["x"] = np.tile(np.arange(nx) * dx, ny)
-        else:
-            d["x"] = np.arange(nx) * dx
-        map_size *= nx
-    if ny > 1:
-        if nx > 1:
-            d["y"] = np.sort(np.tile(np.arange(ny) * dy, nx))
-        else:
-            d["y"] = np.arange(ny) * dy
-        map_size *= ny
-    return d, map_size
+# ---------------------- pytest doctest-modules ---------------------- #
+
+
+@pytest.fixture(autouse=True)
+def doctest_setup_teardown(request):
+    # Setup
+    plt.ioff()  # Interactive plotting off
+    temporary_directory = tempfile.TemporaryDirectory()
+    original_directory = os.getcwd()
+    os.chdir(temporary_directory.name)
+    yield
+
+    # Teardown
+    os.chdir(original_directory)
+    temporary_directory.cleanup()
+    plt.close("all")
+
+
+@pytest.fixture(autouse=True)
+def import_to_namespace(doctest_namespace):
+    DIR_PATH = os.path.dirname(__file__)
+    doctest_namespace["DATA_DIR"] = os.path.join(DIR_PATH, "data/kikuchipy")